requirements.txt
beautifulsoup4
geopy
langchain
langchain-community
langchain-huggingface>=0.0.3
python-dotenv
scikit_learn
scipy
seaborn
SQLAlchemy
streamlit
pandas
polars
numpy
matplotlib
plotly
pydantic
PyJWT==2.8.0
requests
sqlparse
huggingface_hub>=0.17.0